Move generateWordClouds.ts to wordclouds dir
